requirements.txt
PyQt6==6.6.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.1.2
PyJWT==2.8.0
loguru==0.7.2
requests==2.31.0
packaging==23.2
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
Pygments==2.17.2
PyInstaller==6.3.0
